Dictionary Game - BilBakalım
Dictionary Game - Let's Guess: In this game, the system chooses a word from the input words given to the application before. It is a console-based game that guides the user according to the user's predictions, gives hints and calculates the score at the end of the game. It was developed in C++ language. The game is a console-based project written with an object-oriented approach.